requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
cryptography==42.0.0
alembic==1.13.1

# Google Ads & OAuth
google-ads==23.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
pyjwt==2.8.0
passlib==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
httpx==0.26.0
aiohttp==3.9.1
tenacity==8.2.3
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2

# Task Queue (Optional for async tasks)
celery==5.3.6
redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
